pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arango_taxii_server"
version = "0.0.2-pre"
authors = [
  { name="DOGESEC", email="support@dogesec.com" },
]
description = "Arango Taxii Server"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "python-arango>=8.1.3; python_version >= '3.8'",
    "celery>=5.4.0; python_version >= '3.8'",
    "django-cors-headers",
  ]
[project.urls]
Homepage = "https://github.com/muchdogesec/arango_taxii_server"
Issues = "https://github.com/muchdogesec/arango_taxii_server/issues"

[project.optional-dependencies]
full = [
    "stix2arango",
    "redis>=5.1.0b5; python_version >= '3.8'",
    "psycopg2-binary>=2.9.10; python_version >= '3.7'",
]